Add project presentation and change CSMA code to include power amplification for interfering cameras
[Problem]
Previously, we had the assumption that we need not include power amplification for interfering cameras
as they were following different protocol. But, this assumption is not true because we calculcate interference
with the assumption that other cameras follow SEC.

[Solution]
Hence, we need to include power amplification for interfering cameras as well.

[Testing]
Ran the simulation and tested the code.

[Metrics]
Graphs attached in the presentation.

[TODO/Backlog]
Simulate real world setting by avoid the probability matrix completely.

BranchTo: main